requirements.txt
numpy
matplotlib.pyplot
rtlsdr
tqdm
